Clean up and follow official style guide
- Renamed product from `ConcordiumSwiftSdk` to simply `Concordium`. This is the name that will be used for imports, so it should be clear and without redundancy like "swift".
- Removed `./ExampleWallet`. A new one is going to be built from scratch in `./examples`.
- Follow naming conventions from the official [API Design Guidelines](https://www.swift.org/documentation/api-design-guidelines/). I've previously named things according to my personal preferences (i.e. `GrpcNodeClient` instead of `GRPCNodeClient`), but the conventions are clear and we should follow them.
