+++ .gitignore
@@ -1,11 +1,10 @@
 .DS_Store
-/.build/
-/Packages/
-/examples/*/.build/
-/examples/*/Packages/
+.build/
+examples/*/.build/
+examples/*/Packages/
 xcuserdata/
 DerivedData/
-.swiftpm/configuration/registries.json
-.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
+**/.swiftpm/configuration/registries.json
+**/.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
 .netrc
 .idea/
