--- ExampleWallet/ExampleWallet/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,13 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "platform" : "ios",
-      "size" : "1024x1024"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
