--- ExampleWallet/ExampleWallet/ExampleWalletApp.swift
@@ -1,10 +1,0 @@
-import SwiftUI
-
-@main
-struct ExampleWalletApp: App {
-    var body: some Scene {
-        WindowGroup {
-            MainView()
-        }
-    }
-}
