--- ExampleWallet/ExampleWallet/Views/RecoverIdentityView.swift
@@ -1,11 +1,0 @@
-import SwiftUI
-
-struct RecoverIdentityView: View {
-    var body: some View {
-        Text("Nothing to see here yet")
-    }
-}
-
-#Preview {
-    RecoverIdentityView()
-}
