+++ Sources/Concordium/Model/GRPCError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public enum GRPCError: Error, Equatable {
+    case missingBase58CheckVersion(expected: UInt8, actual: UInt8)
+    case missingRequiredValue(String)
+    case unsupportedValue(String)
+    case valueOutOfBounds
+}
