--- Sources/ConcordiumSwiftSdk/Model/Chain.swift
@@ -1,14 +1,0 @@
-import ConcordiumWalletCrypto
-import Foundation
-
-public typealias CryptographicParameters = ConcordiumWalletCrypto.GlobalContext
-
-extension CryptographicParameters {
-    static func fromGrpcType(_ grpc: Concordium_V2_CryptographicParameters) -> Self {
-        .init(
-            onChainCommitmentKeyHex: grpc.onChainCommitmentKey.hex,
-            bulletproofGeneratorsHex: grpc.bulletproofGenerators.hex,
-            genesisString: grpc.genesisString
-        )
-    }
-}
